JOC-1335

+++ sos-webservices/sos-webservices-joc/src/main/java/com/sos/joc/publish/db/DBLayerDeploy.java
@@ -2386,7 +2386,7 @@ public class DBLayerDeploy {
             hql.append(" and type = :type");
             Query<DBItemInventoryReleasedConfiguration> query = getSession().createQuery(hql.toString());
             query.setParameter("name", name);
-            query.setParameter("type", type);
+            query.setParameter("type", type.intValue());
             query.setMaxResults(1);
             return query.getSingleResult();
         } catch (NoResultException e) {
